+++ components/jobs/related-jobs.tsx
@@ -1,38 +1,55 @@
-"use client"
+"use client";
 
-import { useEffect, useState } from "react"
-import Link from "next/link"
-import Image from "next/image"
-import { Briefcase, MapPin, DollarSign } from "lucide-react"
-import { Button } from "@/components/ui/button"
-import { jobService } from "@/services/jobService"
-import type { Job } from "@/types/interfaces"
+import { useEffect, useState } from "react";
+import Link from "next/link";
+import Image from "next/image";
+import {
+  MapPin,
+  Clock,
+  Briefcase,
+  DollarSign,
+  Tags,
+  BarChart,
+} from "lucide-react";
+import { Button } from "@/components/ui/button";
+import { jobService } from "@/services/jobService";
+import type { Job } from "@/types/interfaces";
+import { formatSalary } from "@/utils/formatters";
 
 interface RelatedJobsProps {
-  currentJobId: string
-  category?: string
+  currentJobId: string;
+  category?: string;
 }
 
-export default function RelatedJobs({ currentJobId, category }: RelatedJobsProps) {
-  const [jobs, setJobs] = useState<Job[]>([])
-  const [isLoading, setIsLoading] = useState(true)
+export default function RelatedJobs({
+  currentJobId,
+  category,
+}: RelatedJobsProps) {
+  const [jobs, setJobs] = useState<Job[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
     const fetchRelatedJobs = async () => {
       try {
-        // In a real app, you would fetch related jobs based on category or other criteria
-        const allJobs = await jobService.getJobs({ category })
-        const filteredJobs = allJobs.filter((job) => job.id !== currentJobId).slice(0, 4)
-        setJobs(filteredJobs)
-      } catch (error) {
-        console.error("Failed to fetch related jobs:", error)
-      } finally {
-        setIsLoading(false)
-      }
-    }
+        const res = await jobService.getJobs({
+          category: category ? [category] : undefined,
+        });
+        const allJobs = res?.data || [];
 
-    fetchRelatedJobs()
-  }, [currentJobId, category])
+        const filteredJobs = allJobs
+          .filter((job: Job) => job.id !== currentJobId)
+          .slice(0, 4);
+
+        setJobs(filteredJobs);
+      } catch (error) {
+        console.error("Failed to fetch related jobs:", error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchRelatedJobs();
+  }, [currentJobId, category]);
 
   if (isLoading) {
     return (
@@ -53,13 +70,16 @@ export default function RelatedJobs({ currentJobId, category }: RelatedJobsProps
           </div>
         ))}
       </div>
-    )
+    );
   }
 
   return (
     <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-      {jobs.map((job) => (
-        <div key={job.id} className="border rounded-lg p-4 hover:shadow-md transition">
+      {jobs.map((job, index) => (
+        <div
+          key={job.id ?? `job-${index}`}
+          className="border rounded-lg p-4 hover:shadow-md transition"
+        >
           <div className="flex items-start gap-4">
             <div className="flex-shrink-0">
               <Image
@@ -77,26 +97,49 @@ export default function RelatedJobs({ currentJobId, category }: RelatedJobsProps
                   <p className="text-gray-600 text-sm">{job.company}</p>
                 </div>
                 <div className="text-xs text-gray-500">
-                  {Math.floor(Math.random() * 30) + 1} day{Math.floor(Math.random() * 30) + 1 !== 1 ? "s" : ""} ago
+                  {(job.postedDays || 0)} day{job.postedDays !== 1 ? "s" : ""} ago
                 </div>
               </div>
-              <div className="flex flex-wrap gap-4 mt-3">
+
+              <div className="grid grid-cols-2 md:grid-cols-4 gap-2 mt-4">
                 <div className="flex items-center gap-1 text-sm text-gray-600">
-                  <MapPin className="h-4 w-4 text-gray-400" />
-                  <span>{job.location}</span>
+                  <Tags className="h-4 w-4 text-gray-400" />
+                  <span>{category}</span>
                 </div>
+
                 <div className="flex items-center gap-1 text-sm text-gray-600">
                   <Briefcase className="h-4 w-4 text-gray-400" />
                   <span>{job.type}</span>
                 </div>
                 <div className="flex items-center gap-1 text-sm text-gray-600">
+                  <MapPin className="h-4 w-4 text-gray-400" />
+                  <span>{job.location}</span>
+                </div>
+                <div className="flex items-center gap-1 text-sm text-gray-600">
+                  <BarChart className="h-4 w-4 text-gray-400" />
+                  <span>{job.experienceLevel}</span>
+                </div>
+                <div className="flex items-center gap-1 text-sm text-gray-600">
                   <DollarSign className="h-4 w-4 text-gray-400" />
-                  <span>{job.salary}</span>
+                  <span>{formatSalary(job.salary)}</span>
+                </div>
+                <div className="flex items-center gap-1 text-sm text-gray-600">
+                  <Clock className="h-4 w-4 text-gray-400" />
+                  <span>
+                    Apply by{" "}
+                    {new Date(
+                      Date.now() + 30 * 24 * 60 * 60 * 1000
+                    ).toLocaleDateString()}
+                  </span>
                 </div>
               </div>
             </div>
             <div className="flex-shrink-0">
-              <Button size="sm" className="bg-teal-500 hover:bg-teal-600" asChild>
+              <Button
+                size="sm"
+                className="bg-teal-500 hover:bg-teal-600"
+                asChild
+              >
                 <Link href={`/jobs/${job.id}`}>Job Details</Link>
               </Button>
             </div>
@@ -104,5 +147,5 @@ export default function RelatedJobs({ currentJobId, category }: RelatedJobsProps
         </div>
       ))}
     </div>
-  )
+  );
 }
