+++ templates/contact-information.html
@@ -64,11 +64,7 @@
   <body>
     <div class="container">
       <div class="header">
-        <img
-          src="https://example.com/job-winner-logo.png"
-          alt="Job Winner Logo"
-          class="logo"
-        />
+        <img src="cid:jobwinnerlogo" alt="Logo" style="height:60px" />
       </div>
 
       <div class="content">
