+++ templates/verification-code.html
@@ -85,7 +85,7 @@
 <body>
   <div class="container">
     <div class="header">
-      <img src="/public/logo.svg" alt="Job Winner Logo" class="logo">
+      <img src="cid:jobwinnerlogo" alt="Logo" style="height:60px" />
     </div>
     
     <div class="content">
