+++ templates/verification-contact.html
@@ -65,11 +65,7 @@
   <body>
     <div class="container">
       <div class="header">
-        <img
-          src="https://example.com/job-winner-logo.png"
-          alt="Job Winner Logo"
-          class="logo"
-        />
+        <img src="cid:jobwinnerlogo" alt="Logo" style="height:60px" />
       </div>
 
       <div class="content">
