+++ templates/welcome.html
@@ -93,9 +93,7 @@
 </head>
 <body>
   <div class="container">
-    <div class="header">
-      <img src="/public/logo.svg" alt="Job Winner Logo" class="logo">
-    </div>
+    <img src="cid:jobwinnerlogo" alt="Logo" style="height:60px" />
     
     <div class="content">
       <h1>Chào mừng đến với Job Winner!</h1>
